transforming filterColors jsx to component
