+++ src/assets/css/FilterColors.css
@@ -1,24 +1,33 @@
-.wrap_vali_page{
-    width: 90%;
+.wrap_category_page {
+    width: 1300px;
     margin: 0 auto;
 }
-.vali-item-container{
+
+.category-item-container {
     display: flex;
 }
-.productFilter_container{
+
+.productFilter_container {
     width: 30%;
 }
-.vali_container {
+
+.category_container {
     width: 100%;
     display: flex;
     flex-direction: column;
 }
-.container_crad_vali {
+
+.container_card_category {
     width: 70%;
     display: flex;
     flex-wrap: wrap;
 }
-.title_vali{
+
+.title_category {
     margin-left: 30%;
     font-size: 1.9rem;
+}
+
+.container_card_category .product-card {
+    width: 214px;
 }
